refactor(redux): refactor PR issues
