+++ packages/client/src/store/features/auth/auth.test.ts
@@ -1,8 +1,8 @@
 import MockAdapter from 'axios-mock-adapter';
 
 import { UserDTO } from '../../../api/typings';
-import { buildPath } from '../../../utils/Http';
-import { axios } from '../../../utils/Http/httpClient';
+import { buildPath } from '../../../utils/HTTP';
+import { axios } from '../../../utils/HTTP/httpClient';
 import { store } from '../../store';
 import { API_ENDPOINTS, API_HOST } from './../../../config/constants';
 import { authThunks } from './authThunks';
