+++ packages/client/src/utils/HTTP/HTTP.test.tsx
@@ -1,7 +1,7 @@
 import MockAdapter from 'axios-mock-adapter';
 
 import { buildPath } from './buildPath';
-import { Http } from './Http';
+import { HTTP } from './HTTP';
 import { axios } from './httpClient';
 
 const mock = new MockAdapter(axios);
@@ -12,7 +12,7 @@ describe('Request util', () => {
   test('returns correct data when get request is sent', done => {
     mock.onGet(buildPath(baseUrl, '/users')).reply(200, data);
 
-    Http.get('/users', { baseUrl, params: { page: 2 } }).then(response => {
+    HTTP.get('/users', { baseUrl, params: { page: 2 } }).then(response => {
       expect(response.data).toEqual(data);
       expect(response.status).toEqual(200);
       done();
@@ -22,7 +22,7 @@ describe('Request util', () => {
   test('returns correct data when post request is sent', done => {
     mock.onPost(buildPath(baseUrl, '/users')).reply(200, data);
 
-    Http.post('/users', { baseUrl, data: { name: 'Vasya' } }).then(response => {
+    HTTP.post('/users', { baseUrl, data: { name: 'Vasya' } }).then(response => {
       expect(response.data).toEqual(data);
       expect(response.status).toEqual(200);
       done();
@@ -32,7 +32,7 @@ describe('Request util', () => {
   test('returns correct data when put request is sent', done => {
     mock.onPut(buildPath(baseUrl, '/users')).reply(200, data);
 
-    Http.put('/users', { baseUrl, data: { name: 'Kolya' } }).then(response => {
+    HTTP.put('/users', { baseUrl, data: { name: 'Kolya' } }).then(response => {
       expect(response.data).toEqual(data);
       expect(response.status).toEqual(200);
       done();
@@ -42,7 +42,7 @@ describe('Request util', () => {
   test('returns correct data when delete request is sent', done => {
     mock.onDelete(buildPath(baseUrl, '/users/1')).reply(201, data);
 
-    Http.delete('/users/1', { baseUrl }).then(response => {
+    HTTP.delete('/users/1', { baseUrl }).then(response => {
       expect(response.data).toEqual(data);
       expect(response.status).toEqual(201);
       done();
